requirements.txt
pandas
xlrd
numpy
scikit-learn
matplotlib
nltk
pyyaml


